pyproject.toml
[project]
name = "ms-llm"
version = "0.1.0"
description = "Microservice deployment for LLMs"
authors = [
    {name = "Connor McGuirk",email = "con.mcguirk@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "locust (>=2.37.4,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "llama-cpp-python (>=0.3.9,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.11.10"
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
sphinx = "7.4.7"
furo = "^2024.8.6"
sphinx-autobuild = "^2024.10.3"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# Each test function gets its own fresh event loop. Prevent unfinished tasks interfering with other tests.